+++ org.lxtk.lx4e/src/org/lxtk/lx4e/util/WordFinder.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2021 IBM Corporation and others.
+ * Copyright (c) 2000, 2022 IBM Corporation and others.
  *
  * This program and the accompanying materials are made available under
  * the terms of the Eclipse Public License 2.0 which is available at
@@ -60,8 +60,7 @@ public class WordFinder
      * @param strict the search mode:
      *  <ul>
      *  <li>in strict mode, the offset must fall strictly within the word region;</li>
-     *  <li>in non-strict mode, the offset may immediately follow the word region,
-     *  i.e., the end offset of the word region may equal <code>offset - 1</code>.</li>
+     *  <li>in non-strict mode, the offset may immediately follow the word region</li>
      *  </ul>
      * @return the corresponding word region, or <code>null</code> if none
      */
@@ -73,34 +72,40 @@ public class WordFinder
         try
         {
             int length = document.getLength();
+
+            if (offset > 0 && offset < length && Character.isLowSurrogate(document.getChar(offset))
+                && Character.isHighSurrogate(document.getChar(offset - 1)))
+                --offset; // normalize offset so that it points to the beginning of the surrogate pair
+
             if (!strict && offset > 0
-                && (offset == length || !isWordPart(document.getChar(offset))))
-                --offset;
+                && (offset == length || !isWordPart(codePointAt(document, offset))))
+                offset -= Character.charCount(codePointBefore(document, offset));
 
             int pos = offset;
             do
             {
-                if (!isWordPart(document.getChar(pos)))
+                int c = codePointAt(document, pos);
+                if (!isWordPart(c))
                     break;
-                --pos;
+                pos += Character.charCount(c);
             }
-            while (pos >= 0);
+            while (pos < length);
 
             if (pos == offset)
                 return null;
 
-            start = pos + 1;
+            end = pos;
 
             pos = offset;
-            do
+            while (pos > 0)
             {
-                if (!isWordPart(document.getChar(pos)))
+                int c = codePointBefore(document, pos);
+                if (!isWordPart(c))
                     break;
-                ++pos;
+                pos -= Character.charCount(c);
             }
-            while (pos < length);
 
-            end = pos;
+            start = pos;
         }
         catch (BadLocationException e)
         {
@@ -113,12 +118,36 @@ public class WordFinder
     /**
      * Determines whether the given character may be part of a word.
      *
-     * @param ch a character
+     * @param c a character (Unicode code point)
      * @return <code>true</code> if the given character may be part of a word,
      *  and <code>false</code> otherwise
      */
-    protected boolean isWordPart(char ch)
+    protected boolean isWordPart(int c)
     {
-        return Character.isUnicodeIdentifierPart(ch);
+        return Character.isUnicodeIdentifierPart(c);
+    }
+
+    private static int codePointAt(IDocument document, int offset) throws BadLocationException
+    {
+        char ch = document.getChar(offset);
+        if (Character.isHighSurrogate(ch) && offset < document.getLength() - 1)
+        {
+            char ch2 = document.getChar(offset + 1);
+            if (Character.isLowSurrogate(ch2))
+                return Character.toCodePoint(ch, ch2);
+        }
+        return ch;
+    }
+
+    private static int codePointBefore(IDocument document, int offset) throws BadLocationException
+    {
+        char ch = document.getChar(offset - 1);
+        if (Character.isLowSurrogate(ch) && offset > 1)
+        {
+            char ch2 = document.getChar(offset - 2);
+            if (Character.isHighSurrogate(ch2))
+                return Character.toCodePoint(ch2, ch);
+        }
+        return ch;
     }
 }